Added segmentation tests.

+++ MOIMS_TESTBED_MALSPP_FRAMEWORK/src/main/java/org/ccsds/moims/mo/malspp/test/patterns/MalSppPatternTest.java
@@ -235,7 +235,7 @@ public class MalSppPatternTest extends PatternTest {
     return spacePacketCheck.readDestinationId();
   }
   
-  public int segmentCounterIs() {
+  public long segmentCounterIs() {
     return spacePacketCheck.segmentCounterIs();
   }
